correction on model prediction calculations-algorithms.md
In the original code the y_test series was passed to the model to make predictions. Apart from being conceptually wrong (we are giving the data to predict to the model as input for predictions) this also would raise an error because the model expects data with the same number of columns as X_train.